+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Nedenfor ses dato for release og beskrivelse af opgaver som er implementeret.
 * Opdaterede installationsguide.
 * Applied Maestro notification patch
   (<https://github.com/itk-dev/os2forms_selvbetjening/pull/270>)
+* Opdaterede 4xx page logout link
 
 ## [2.7.3] 2024-01-16